Cleanup types package before going public.
