+++ pkg/types/genargs.go
@@ -1,18 +1,5 @@
-/*
-Copyright 2019 The Kubernetes Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
+// Copyright 2019 The Kubernetes Authors.
+// SPDX-License-Identifier: Apache-2.0
 
 package types
 
@@ -21,7 +8,7 @@ import (
 	"strings"
 )
 
-// GenArgs contains both generator args and options
+// GenArgs contains both GeneratorArgs and GeneratorOptions.
 type GenArgs struct {
 	args *GeneratorArgs
 	opts *GeneratorOptions
